requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
asgiref==3.11.1
backoff==2.2.1
bcrypt==5.0.0
build==1.4.3
camel-converter==4.0.1
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
chroma-hnswlib==0.7.6
chromadb==0.5.5
click==8.1.8
coloredlogs==15.0.1
cryptography==46.0.7
distro==1.9.0
durationpy==0.10
exceptiongroup==1.3.1
fastapi==0.128.8
filelock==3.19.1
flatbuffers==25.12.19
fsspec==2025.10.0
google-ai-generativelanguage==0.6.15
google-api-core==2.30.3
google-api-python-client==2.194.0
google-auth==2.49.2
google-auth-httplib2==0.3.1
google-generativeai==0.8.6
googleapis-common-protos==1.74.0
grpcio==1.80.0
grpcio-status==1.71.2
h11==0.16.0
hf-xet==1.4.3
httpcore==1.0.9
httplib2==0.31.2
httptools==0.7.1
httpx==0.28.1
huggingface_hub==0.36.2
humanfriendly==10.0
idna==3.11
importlib_metadata==8.7.1
importlib_resources==6.5.2
Jinja2==3.1.6
joblib==1.5.3
kubernetes==35.0.0
markdown-it-py==3.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
meilisearch==0.38.0
mmh3==5.2.0
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.4
oauthlib==3.3.1
onnxruntime==1.19.2
openai==1.37.1
opentelemetry-api==1.41.0
opentelemetry-exporter-otlp-proto-common==1.41.0
opentelemetry-exporter-otlp-proto-grpc==1.41.0
opentelemetry-instrumentation==0.62b0
opentelemetry-instrumentation-asgi==0.62b0
opentelemetry-instrumentation-fastapi==0.62b0
opentelemetry-proto==1.41.0
opentelemetry-sdk==1.41.0
opentelemetry-semantic-conventions==0.62b0
opentelemetry-util-http==0.62b0
orjson==3.11.5
overrides==7.7.0
packaging==26.1
pillow==11.3.0
posthog==6.9.3
proto-plus==1.27.2
protobuf==5.29.6
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.13.1
pydantic_core==2.46.1
Pygments==2.20.0
pyparsing==3.3.2
pypdf==4.3.0
PyPika==0.51.1
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
PyYAML==6.0.3
regex==2026.1.15
requests==2.32.5
requests-oauthlib==2.0.0
rich==15.0.0
safetensors==0.7.0
scikit-learn==1.6.1
scipy==1.13.1
sentence-transformers==3.0.1
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
starlette==0.49.3
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tokenizers==0.22.2
tomli==2.4.1
torch==2.8.0
tqdm==4.67.3
transformers==4.57.6
typer==0.23.2
typing-inspection==0.4.2
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.6.3
uvicorn==0.39.0
uvloop==0.22.1
watchdog==4.0.1
watchfiles==1.1.1
websocket-client==1.9.0
websockets==15.0.1
wrapt==2.1.2
zipp==3.23.1
